requirements.txt
python-dotenv
pandas
numpy
pymango
certifi
pymongo[srv]>=4.1.0
scikit-learn
pyaml
mlflow


##-e .